requirements.txt
accelerate==1.12.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
diffusers==0.36.0
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
peft==0.18.0
pillow==11.3.0
psutil==7.1.3
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
setuptools==70.2.0
shellingham==1.5.4
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1+cu128
torchaudio==2.9.1+cu128
torchvision==0.24.1+cu128
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typer-slim==0.20.0
typing_extensions==4.15.0
urllib3==2.6.1
zipp==3.23.0
